Refactored mutect2/run wrapper, and added unit tests to class ParallelMutect2Wrapper.
